Исправляем тип возвращаемого значения multiply. Убираем не нужные exit;

+++ Multiplyer/loader.h
@@ -4,7 +4,7 @@
 #include <Windows.h>
 #include <errhandlingapi.h>
 
-typedef std::vector<int>& (*MultiplyFuncPointer)(std::vector<int>, int, double);
+typedef std::vector<double>& (*MultiplyFuncPointer)(std::vector<int>, int, double);
 
 MultiplyFuncPointer loadFunctionFromDLL()
 {
